+++ resources/views/themes/zeus/post.blade.php
@@ -9,8 +9,8 @@
         </li>
     </x-slot>
 
-    @if(!$post->getMedia('posts')->isEmpty())
-        <img src="{{ $post->getFirstMediaUrl('posts') }}" class="my-10 w-full h-full shadow-md rounded-[2rem] rounded-bl-none z-0 object-cover"/>
+    @if($post->image() !== null)
+        <img alt="{{ $post->title }}" src="{{ $post->image() }}" class="my-10 w-full h-full shadow-md rounded-[2rem] rounded-bl-none z-0 object-cover"/>
     @endif
 
     <div class="bg-white dark:bg-gray-800 rounded-[2rem] rounded-tl-none shadow-md px-10 pb-6">
@@ -18,7 +18,7 @@
             <span class="font-light text-gray-600 dark:text-gray-200">{{ optional($post->published_at)->diffForHumans() ?? '' }}</span>
             <div>
                 @unless ($post->tags->isEmpty())
-                    @each($theme.'.partial.tag', $post->tags->where('type','category'), 'tag')
+                    @each($theme.'.partial.category', $post->tags->where('type','category'), 'category')
                 @endunless
             </div>
         </div>
@@ -31,6 +31,13 @@
                 <p class="mt-2 text-gray-600 dark:text-gray-200">
                     {{ $post->description ?? '' }}
                 </p>
+                <div>
+                    @unless ($post->tags->isEmpty())
+                        @foreach($post->tags->where('type','tag') as $tag)
+                            @include($theme.'.partial.tag')
+                        @endforeach
+                    @endunless
+                </div>
             </div>
             <a href="#" class="flex items-center gap-2">
                 <img src="{{ \Filament\Facades\Filament::getUserAvatarUrl($post->author) }}" alt="avatar" class="object-cover w-10 h-10 rounded-full sm:block">
@@ -39,17 +46,19 @@
         </div>
 
         <div class="mt-6 lg:mt-12 prose dark:prose-invert max-w-none">
-            {!! html_entity_decode($post->content) !!}
+            {!! $post->getContent() !!}
         </div>
     </div>
 
-    <div class="py-6 flex flex-col mt-4 gap-4">
-        <h1 class="text-xl font-bold text-gray-700 dark:text-gray-100 md:text-2xl">{{ __('Related Posts') }}</h1>
+    @if($related->isNotEmpty())
+        <div class="py-6 flex flex-col mt-4 gap-4">
+            <h1 class="text-xl font-bold text-gray-700 dark:text-gray-100 md:text-2xl">{{ __('Related Posts') }}</h1>
 
-        <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
-            @foreach($related as $post)
-                @include($theme.'.partial.related')
-            @endforeach
+            <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
+                @foreach($related as $post)
+                    @include($theme.'.partial.related')
+                @endforeach
+            </div>
         </div>
-    </div>
+    @endif
 </div>