+++ src/Filament/Resources/LibraryResource/Pages/CreateLibrary.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\LibraryResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use LaraZeus\Sky\Filament\Resources\LibraryResource;
+
+class CreateLibrary extends CreateRecord
+{
+    use CreateRecord\Concerns\Translatable;
+
+    protected static string $resource = LibraryResource::class;
+}