+++ src/Filament/Resources/LibraryResource/Pages/EditLibrary.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Sky\Filament\Resources\LibraryResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use LaraZeus\Sky\Filament\Resources\LibraryResource;
+
+class EditLibrary extends EditRecord
+{
+    use EditRecord\Concerns\Translatable;
+
+    protected static string $resource = LibraryResource::class;
+}