+++ src/Filament/Resources/PageResource/Pages/ListPosts.php
@@ -4,8 +4,8 @@ namespace LaraZeus\Sky\Filament\Resources\PageResource\Pages;
 
 use Filament\Resources\Pages\ListRecords;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\SoftDeletingScope;
 use LaraZeus\Sky\Filament\Resources\PageResource;
-use LaraZeus\Sky\Models\Post;
 
 class ListPosts extends ListRecords
 {
@@ -15,6 +15,10 @@ class ListPosts extends ListRecords
 
     protected function getTableQuery(): Builder
     {
-        return Post::where('post_type', 'page');
+        return config('zeus-sky.models.post')::query()
+            ->where('post_type', 'page')
+            ->withoutGlobalScopes([
+                SoftDeletingScope::class,
+            ]);
     }
 }