+++ src/Http/Livewire/Post.php
@@ -2,7 +2,6 @@
 
 namespace LaraZeus\Sky\Http\Livewire;
 
-use LaraZeus\Sky\Models\Post as postModel;
 use Livewire\Component;
 
 class Post extends Component
@@ -11,15 +10,31 @@ class Post extends Component
 
     public function mount($slug)
     {
-        $this->post = postModel::Published()->where('slug', $slug)->firstOrFail();
+        $this->post = config('zeus-sky.models.post')::where('slug', $slug)->firstOrFail();
     }
 
     public function render()
     {
-        if (! $this->post->getMedia('posts')->isEmpty()) {
-            seo()->image($this->post->getFirstMediaUrl('posts'));
+        $this->setSeo();
+
+        if ($this->post->status !== 'publish' && ! $this->post->require_password) {
+            abort_if(! auth()->check(), 404);
+            abort_if($this->post->user_id !== auth()->user()->id, 401);
         }
 
+        if ($this->post->require_password && ! session()->has($this->post->slug . '-' . $this->post->password)) {
+            return view(app('theme') . '.partial.password-form')
+                ->layout(config('zeus-sky.layout'));
+        }
+
+        return view(app('theme') . '.post')
+            ->with('post', $this->post)
+            ->with('related', config('zeus-sky.models.post')::related($this->post)->take(4)->get())
+            ->layout(config('zeus-sky.layout'));
+    }
+
+    public function setSeo(): void
+    {
         seo()
             ->title($this->post->title)
             ->description(($this->post->description ?? '') . ' ' . config('zeus-sky.site_description', 'Laravel'))
@@ -29,9 +44,8 @@ class Post extends Component
             ->withUrl()
             ->twitter();
 
-        return view(app('theme') . '.post')
-            ->with('post', $this->post)
-            ->with('related', postModel::related($this->post)->take(4)->get())
-            ->layout(config('zeus-sky.layout'));
+        if (! $this->post->getMedia('posts')->isEmpty()) {
+            seo()->image($this->post->getFirstMediaUrl('posts'));
+        }
     }
 }