+++ src/Models/Post.php
@@ -3,8 +3,11 @@
 namespace LaraZeus\Sky\Models;
 
 use Database\Factories\PostFactory;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Blade;
 use Spatie\MediaLibrary\HasMedia;
@@ -20,9 +23,18 @@ use Spatie\Translatable\HasTranslations;
  * @property string $status
  * @property string $slug
  * @property string $post_type
+ * @property int $user_id
+ * @property bool $require_password
+ * @property string $password
+ *
+ * @method Builder|static sticky()
+ * @method Builder|static published()
+ * @method Builder|static query()
+ * @method Builder|static withAnyTags()
  */
 class Post extends Model implements HasMedia
 {
+    use SoftDeletes;
     use HasFactory;
     use HasTags;
     use InteractsWithMedia;
@@ -68,7 +80,7 @@ class Post extends Model implements HasMedia
 
     public function statusDesc(): string
     {
-        $PostStatus = PostStatus::where('name', $this->status)->first();
+        $PostStatus = config('zeus-sky.models.postStatus')::where('name', $this->status)->first();
         $icon = Blade::render('@svg("' . $PostStatus->icon . '","w-4 h-4 inline-flex")');
 
         return "<span title='" . __('post status') . "' class='$PostStatus->class'> " . $icon . " {$PostStatus->label}</span>";
@@ -82,5 +94,26 @@ class Post extends Model implements HasMedia
     public function parent()
     {
         return $this->belongsTo(self::class, 'parent_id', 'id');
+    }
+
+    public function image()
+    {
+        if (! $this->getMedia('posts')->isEmpty()) {
+            return $this->getFirstMediaUrl('posts');
+        } else {
+            return $this->featured_image ?? config('zeus-sky.default_featured_image', null);
+        }
+    }
+
+    protected function requirePassword(): Attribute
+    {
+        return Attribute::make(
+            get: fn () => $this->status === 'private' && $this->password !== null,
+        );
+    }
+
+    public function getContent(): string
+    {
+        return config('zeus-sky.editor')::render($this->content);
     }
 }